refactor: extract App initialization into useAppInit composable
Extracted the bloated global startup logic from `web/src/App.vue` (including fullscreen synchronization, Grok API gateway wiring, and update checks) into a new `web/src/composables/useAppInit.js` file.

Fixed the unused variable ESLint failure inside the newly created composable.
