+++ web/src/App.vue
@@ -101,7 +101,7 @@
 </template>
 
 <script setup>
-import { onMounted, onUnmounted, ref, watch, computed } from 'vue'
+import { ref, watch, computed } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { ElMessage } from 'element-plus'
 import AppLogo from './components/AppLogo.vue'
@@ -111,28 +111,29 @@ import {
   furnaceSpriteState,
   furnaceWorkspaceOpen,
   grokProbe,
-  setGrokConfigured,
-  setFurnaceGrokGate,
   grokCanRun,
   grokSetupNeeded,
 } from './composables/furnaceUi.js'
 import { FURNACE_DISPLAY_NAME } from '@acw/shared'
 import { api } from './api'
-import {
-  exitFullscreen,
-  fullscreenElement,
-  requestFullscreen,
-} from './composables/fullscreen'
+import { exitFullscreen, fullscreenElement, requestFullscreen } from './composables/fullscreen'
+import { useAppInit } from './composables/useAppInit.js'
 
 const route = useRoute()
 const router = useRouter()
 const nav = ref(route.path.startsWith('/settings') ? 'settings' : 'workbench')
 const appRoot = ref(null)
-const isFullscreen = ref(false)
-const grokGuideOpen = ref(false)
-const grokGuideStatus = ref({})
-const grokExampleToml = ref('')
-const grokCanContinue = computed(() => grokCanRun(grokGuideStatus.value))
+
+const {
+  isFullscreen,
+  grokGuideOpen,
+  grokGuideStatus,
+  grokExampleToml,
+  grokCanContinue,
+  refreshGrokGate,
+  ensureFurnaceGrokWired,
+} = useAppInit()
+
 const furnaceTitle = computed(() => {
   const probe = grokProbe.value
   const gaps = probe?.gaps || []
@@ -149,10 +150,6 @@ const furnaceTitle = computed(() => {
 })
 
 const isDocsRoute = computed(() => route.path.startsWith('/docs'))
-
-function syncFullscreenState() {
-  isFullscreen.value = !!fullscreenElement()
-}
 
 async function toggleWorkbenchFullscreen() {
   const ok = fullscreenElement()
@@ -177,39 +174,9 @@ function goWorkbench() {
   router.push('/workbench')
 }
 
-async function refreshGrokGate() {
-  try {
-    const [s, probe] = await Promise.all([api.appSettings.get(), api.grok.status()])
-    grokGuideStatus.value = probe
-    grokExampleToml.value = probe.exampleToml || ''
-    setFurnaceGrokGate({ probe })
-    return { s, probe }
-  } catch {
-    setGrokConfigured(false)
-    return { s: null, probe: null }
-  }
-}
-
 function openFurnaceSession() {
   const path = route.path.startsWith('/workbench') ? route.path : '/workbench'
   router.push({ path, query: { ...route.query, furnace: '1' } })
-}
-
-/** 本机已能跑 Grok 时，把熔炉成员接到 grok 命令，否则只开聊天回声 */
-async function ensureFurnaceGrokWired(s, probe) {
-  if (!grokCanRun(probe) || s?.grok?.configured) return s
-  try {
-    const next = await api.appSettings.update({
-      grok: {
-        command: s?.grok?.command || probe?.command || 'grok',
-        configured: true,
-      },
-    })
-    setFurnaceGrokGate({ probe })
-    return next
-  } catch {
-    return s
-  }
 }
 
 async function onFurnaceClick() {
@@ -229,22 +196,6 @@ async function onFurnaceClick() {
   openFurnaceSession()
 }
 
-// 4.7.0 启动时检查更新（静默，仅发现新版本时轻提示）
-async function startupCheckUpdate() {
-  try {
-    const s = await api.appSettings.get()
-    if (s.updateCheck?.startup !== true) return
-    const r = await api.update.check()
-    if (!r.checked || !r.hasUpdate) return
-    ElMessage.info({
-      message: `发现新版本 v${r.latest}，前往设置 → 关于查看详情`,
-      duration: 8000,
-    })
-  } catch {
-    // 静默失败，不打扰用户
-  }
-}
-
 async function openFurnaceAnyway() {
   grokGuideOpen.value = false
   const { s, probe } = await refreshGrokGate()
@@ -261,17 +212,6 @@ function goGrokSettings() {
   grokGuideOpen.value = false
   router.push('/settings/prefs')
 }
-
-onMounted(() => {
-  document.addEventListener('fullscreenchange', syncFullscreenState)
-  syncFullscreenState()
-  refreshGrokGate()
-  startupCheckUpdate()
-})
-
-onUnmounted(() => {
-  document.removeEventListener('fullscreenchange', syncFullscreenState)
-})
 </script>
 
 <style scoped>
